pyproject.toml
[project]
name = "chimera"
version = "0.1.0"
description = "Autonomous multi-model orchestration system. Six composable LangGraph patterns unified under a Kabbalistic architecture where intent descends through balanced forces into running code."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-core>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2",
    "langgraph>=0.4",
    "langgraph-checkpoint-sqlite>=2",
    "aiosqlite>=0.20",
    "mcp[cli]>=1.0",
    "python-dotenv>=1",
    "pyyaml>=6",
    "pydantic>=2",
    "anyio>=4",
    "psutil>=5",
]

[project.optional-dependencies]
# `chimera[monitor]` — installs LangGraph monitor backend deps.
# FastAPI is already pulled in transitively via mcp[cli]; pinned here for safety
# in case mcp drops it. qdrant-client added in Phase 2 when the Qdrant view ships.
monitor = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "psycopg[binary]>=3.1",
    "tree-sitter>=0.21",
    "tree-sitter-python>=0.21",
]

[project.scripts]
chimera = "chimera:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
    "pytest>=8",
]
